Me mori de risa cuando me di cuenta que el poisson devuelve doubles, cambie todo que referido a tiempo de int a double.

+++ src/events/Evento.java
@@ -15,7 +15,7 @@ public abstract class Evento
         this.auto = auto;
     }
     
-    public abstract int getHorario();
+    public abstract double getHorario();
     
     public Auto getAuto()
     {
